pyproject.toml
[tool.poetry]
name = "ifs-sequel"
version = "0.1.0"
description = ""
authors = ["Jeffrey Katz <JeffreyEricKatz@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
langchain-openai = "^0.1.21"
langchain = "^0.2.14"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.12"
chainlit = "^1.1.402"
langchain_core = "*"
asyncio  = "*"
flask = "*"
flask-cors = "*"
hypercorn = "*"
quart = "*"
quart-cors = "*"
langsmith = "^0.1.121"
rapidfuzz = "*"
deepeval = "*"
pytest = "^8.3.3"
pytest-asyncio = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

